+++ src/calmjs/parse/unparsers/es5.py
@@ -24,6 +24,8 @@ from calmjs.parse.ruletypes import (
 )
 from calmjs.parse.ruletypes import (
     Attr,
+    CommentsAttr,
+    Iter,
     Text,
     Optional,
     JoinAttr,
@@ -34,6 +36,8 @@ from calmjs.parse.ruletypes import (
 from calmjs.parse.ruletypes import (
     Literal,
     Declare,
+    LineComment,
+    BlockComment,
     Resolve,
     ResolveFuncName,
 )
@@ -45,6 +49,7 @@ from calmjs.parse.unparsers.base import BaseUnparser
 from calmjs.parse import rules
 
 value = (
+    CommentsAttr(),
     Attr('value'),
 )
 
@@ -55,6 +60,7 @@ definitions = {
         OptionalNewline,
     ),
     'Block': (
+        CommentsAttr(),
         OpenBlock,
         Indent, Newline,
         children_newline,
@@ -62,26 +68,35 @@ definitions = {
         CloseBlock,
     ),
     'VarStatement': (
+        CommentsAttr(),
         Text(value='var'), RequiredSpace, children_comma, EndStatement,
     ),
     'VarDecl': (
+        CommentsAttr(),
         Attr(Declare('identifier')), Optional('initializer', (
             Space, Operator(value='='), Space, Attr('initializer'),),),
     ),
     'VarDeclNoIn': (
+        CommentsAttr(),
         Text(value='var '), Attr(Declare('identifier')),
         Optional('initializer', (
             Space, Operator(value='='), Space, Attr('initializer'),),),
     ),
     'GroupingOp': (
+        CommentsAttr(),
         Text(value='('), Attr('expr'), Text(value=')'),
     ),
-    'Identifier': (Attr(Resolve()),),
+    'Identifier': (
+        CommentsAttr(),
+        Attr(Resolve()),
+    ),
     'PropIdentifier': value,
     'Assign': (
+        CommentsAttr(),
         Attr('left'), OptionalSpace, Attr('op'), Space, Attr('right'),
     ),
     'GetPropAssign': (
+        CommentsAttr(),
         Text(value='get'), RequiredSpace, Attr('prop_name'),
         PushScope,
         Text(value='('), Text(value=')'), Space,
@@ -93,6 +108,7 @@ definitions = {
         PopScope,
     ),
     'SetPropAssign': (
+        CommentsAttr(),
         Text(value='set'), RequiredSpace, Attr('prop_name'), Text(value='('),
         PushScope,
         Attr(Declare('parameter')), Text(value=')'), Space,
@@ -105,10 +121,15 @@ definitions = {
     ),
     'Number': value,
     'Comma': (
+        CommentsAttr(),
         Attr('left'), Text(value=','), Space, Attr('right'),
     ),
-    'EmptyStatement': (EndStatement,),
+    'EmptyStatement': (
+        CommentsAttr(),
+        EndStatement,
+    ),
     'If': (
+        CommentsAttr(),
         Text(value='if'), Space,
         Text(value='('), Attr('predicate'), Text(value=')'), Space,
         Attr('consequent'),
@@ -117,6 +138,7 @@ definitions = {
     ),
     'Boolean': value,
     'For': (
+        CommentsAttr(),
         Text(value='for'), Space, Text(value='('),
         Attr('init'),
         OptionalSpace, Attr('cond'),
@@ -124,52 +146,65 @@ definitions = {
         Text(value=')'), Space, Attr('statement'),
     ),
     'ForIn': (
+        CommentsAttr(),
         Text(value='for'), Space, Text(value='('),
         Attr('item'),
         RequiredSpace, Text(value='in'), Space,
         Attr('iterable'), Text(value=')'), Space, Attr('statement'),
     ),
     'BinOp': (
+        CommentsAttr(),
         Attr('left'), Space, Operator(attr='op'), Space, Attr('right'),
     ),
     'UnaryExpr': (
+        CommentsAttr(),
         Operator(attr='op'), OptionalSpace, Attr('value'),
     ),
     'PostfixExpr': (
+        CommentsAttr(),
         Operator(attr='value'), Attr('op'),
     ),
     'ExprStatement': (
+        CommentsAttr(),
         Attr('expr'), EndStatement,
     ),
     'DoWhile': (
+        CommentsAttr(),
         Text(value='do'), Space, Attr('statement'), Space,
         Text(value='while'), Space, Text(value='('),
         Attr('predicate'), Text(value=')'), EndStatement,
     ),
     'While': (
+        CommentsAttr(),
         Text(value='while'), Space,
         Text(value='('), Attr('predicate'), Text(value=')'), OptionalSpace,
         Attr('statement'),
     ),
     'Null': (
+        CommentsAttr(),
         Text(value='null'),
     ),
     'String': (
+        CommentsAttr(),
         Attr(Literal()),
     ),
     'Continue': (
+        CommentsAttr(),
         Text(value='continue'), Optional('identifier', (
             RequiredSpace, Attr(attr='identifier'))), EndStatement,
     ),
     'Break': (
+        CommentsAttr(),
         Text(value='break'), Optional('identifier', (
             RequiredSpace, Attr(attr='identifier'),)), EndStatement,
     ),
     'Return': (
+        CommentsAttr(),
         Text(value='return'), Optional('expr', (
             Space, Attr(attr='expr'))), EndStatement,
     ),
     'With': (
+        CommentsAttr(),
         Text(value='with'), Space,
         # should _really_ have a token for logging a warning
         # https://developer.mozilla.org/en-US/docs/Web/JavaScript/
@@ -178,9 +213,11 @@ definitions = {
         Attr('statement'),
     ),
     'Label': (
+        CommentsAttr(),
         Attr('identifier'), Text(value=':'), Space, Attr('statement'),
     ),
     'Switch': (
+        CommentsAttr(),
         Text(value='switch'), Space,
         Text(value='('), Attr('expr'), Text(value=')'), Space,
         Attr('case_block'),
@@ -193,29 +230,35 @@ definitions = {
         CloseBlock,
     ),
     'Case': (
+        CommentsAttr(),
         Text(value='case'), Space, Attr('expr'), Text(value=':'),
         Indent, Newline,
         JoinAttr('elements', value=(Newline,)),
         Dedent,
     ),
     'Default': (
+        CommentsAttr(),
         Text(value='default'), Text(value=':'),
         Indent, Newline,
         JoinAttr('elements', value=(Newline,)),
         Dedent,
     ),
     'Throw': (
+        CommentsAttr(),
         Text(value='throw'), Space, Attr('expr'), EndStatement,
     ),
     'Debugger': (
+        CommentsAttr(),
         Text(value='debugger'), EndStatement,
     ),
     'Try': (
+        CommentsAttr(),
         Text(value='try'), Space, Attr('statements'),
         Optional('catch', (Newline, Attr('catch'),)),
         Optional('fin', (Newline, Attr('fin'),)),
     ),
     'Catch': (
+        CommentsAttr(),
         Text(value='catch'), Space,
         PushCatch,
         Text(value='('), Attr('identifier'), Text(value=')'), Space,
@@ -223,9 +266,11 @@ definitions = {
         PopCatch,
     ),
     'Finally': (
+        CommentsAttr(),
         Text(value='finally'), Space, Attr('elements'),
     ),
     'FuncDecl': (
+        CommentsAttr(),
         Text(value='function'), Optional('identifier', (RequiredSpace,)),
         Attr(Declare('identifier')), Text(value='('),
         PushScope, Optional('identifier', (ResolveFuncName,)),
@@ -239,6 +284,7 @@ definitions = {
         PopScope,
     ),
     'FuncExpr': (
+        CommentsAttr(),
         Text(value='function'), Optional('identifier', (RequiredSpace,)),
         Attr(Declare('identifier')), Text(value='('),
         PushScope, Optional('identifier', (ResolveFuncName,)),
@@ -252,29 +298,36 @@ definitions = {
         PopScope,
     ),
     'Conditional': (
+        CommentsAttr(),
         Attr('predicate'), Space, Text(value='?'), Space,
         Attr('consequent'), Space, Text(value=':'), Space,
         Attr('alternative'),
     ),
     'Regex': value,
     'NewExpr': (
+        CommentsAttr(),
         Text(value='new'), RequiredSpace, Attr('identifier'), Attr('args'),
     ),
     'DotAccessor': (
+        CommentsAttr(),
         Attr('node'), Text(value='.'), Attr('identifier'),
     ),
     'BracketAccessor': (
+        CommentsAttr(),
         Attr('node'), Text(value='['), Attr('expr'), Text(value=']'),
     ),
     'FunctionCall': (
+        CommentsAttr(),
         Attr('identifier'), Attr('args'),
     ),
     'Arguments': (
+        CommentsAttr(),
         Text(value='('),
         JoinAttr('items', value=(Text(value=','), Space)),
         Text(value=')'),
     ),
     'Object': (
+        CommentsAttr(),
         Text(value='{'),
         Optional('properties', (
             Indent, Newline,
@@ -284,15 +337,27 @@ definitions = {
         Text(value='}'),
     ),
     'Array': (
+        CommentsAttr(),
         Text(value='['),
         ElisionJoinAttr('items', value=(Space,)),
         Text(value=']'),
     ),
     'Elision': (
+        CommentsAttr(),
         ElisionToken(attr='value', value=','),
     ),
     'This': (
+        CommentsAttr(),
         Text(value='this'),
+    ),
+    'Comments': (
+        JoinAttr(Iter(), value=()),
+    ),
+    'LineComment': (
+        Attr(LineComment()), Newline,
+    ),
+    'BlockComment': (
+        Attr(BlockComment()), Newline,
     ),
 }